Move test_em_asm_parameter_pack out to a file
