+++ tests/core/test_em_asm_parameter_pack.out
@@ -1,0 +1,3 @@
+1
+1,2
+1,2,3
